moving analysis visitors into analysis module.
