+++ src/bloqade/codegen/emulator_ir.py
@@ -14,7 +14,7 @@ import bloqade.ir.control.pulse as pulse
 import bloqade.ir.control.waveform as waveform
 import bloqade.ir.control.field as field
 import bloqade.ir as ir
-from bloqade.codegen.common.is_hyperfine import IsHyperfineSequence
+from bloqade.ir.analysis.is_hyperfine import IsHyperfineSequence
 from bloqade.emulate.ir.atom_type import ThreeLevelAtom, TwoLevelAtom
 from bloqade.emulate.ir.emulator import (
     DetuningOperatorData,
