+++ src/bloqade/ir/analysis/is_constant_sequence.py
