+++ src/bloqade/ir/analysis/is_hyperfine.py
